refactor: move veil handling for metric popups in popup component
